backend/requirements.txt
fastapi
uvicorn[standard]
motor
beanie
python-dotenv
pydantic
colorama
python-multipart
passlib[bcrypt]
python-jose[cryptography]
email-validator
